requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
pymongo>=4.7.0
dnspython>=2.6.0
requests>=2.32.0
feedparser>=6.0.11
jinja2>=3.1.4
beautifulsoup4>=4.12.0
